Reorganize and implement GET /scene/:id
Group all of the Constellations dataclasses in
`wwt_api_client.constellations.data`, which avoids problems about
deciding where to put various classes and avoiding circular
dependencies.

Document the new `GET /scene/:id` endpoint.

Create a new `wwt_api_client.constellations.scenes`, analogous to
`wwt_api_client.constellations.handles`, with a client class specific to
individual scenes. Add a `get()` method to get scene info.
