+++ docs/api/wwt_api_client.constellations.data.SceneContent.rst
@@ -1,0 +1,25 @@
+SceneContent
+============
+
+.. currentmodule:: wwt_api_client.constellations.data
+
+.. autoclass:: SceneContent
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~SceneContent.from_dict
+      ~SceneContent.from_json
+      ~SceneContent.schema
+      ~SceneContent.to_dict
+      ~SceneContent.to_json
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: from_dict
+   .. automethod:: from_json
+   .. automethod:: schema
+   .. automethod:: to_dict
+   .. automethod:: to_json
