+++ docs/api/wwt_api_client.constellations.data.rst
@@ -1,0 +1,3 @@
+.. automodapi:: wwt_api_client.constellations.data
+   :no-inheritance-diagram:
+   :no-inherited-members:
