+++ docs/api/wwt_api_client.constellations.scenes.GetSceneResponse.rst
@@ -1,0 +1,25 @@
+GetSceneResponse
+================
+
+.. currentmodule:: wwt_api_client.constellations.scenes
+
+.. autoclass:: GetSceneResponse
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~GetSceneResponse.from_dict
+      ~GetSceneResponse.from_json
+      ~GetSceneResponse.schema
+      ~GetSceneResponse.to_dict
+      ~GetSceneResponse.to_json
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: from_dict
+   .. automethod:: from_json
+   .. automethod:: schema
+   .. automethod:: to_dict
+   .. automethod:: to_json
