+++ docs/api/wwt_api_client.constellations.scenes.SceneClient.rst
@@ -1,0 +1,17 @@
+SceneClient
+===========
+
+.. currentmodule:: wwt_api_client.constellations.scenes
+
+.. autoclass:: SceneClient
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~SceneClient.get
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get
