+++ docs/api/wwt_api_client.constellations.scenes.rst
@@ -1,0 +1,3 @@
+.. automodapi:: wwt_api_client.constellations.scenes
+   :no-inheritance-diagram:
+   :no-inherited-members:
