+++ wwt_api_client/constellations/handles.py
@@ -10,49 +10,32 @@ Handles are publicly visible "user" or "channel" names in Constellations.
 from dataclasses import dataclass
 from dataclasses_json import dataclass_json
 import math
-from typing import List, Optional
+from typing import Optional
 import urllib.parse
 
 from wwt_data_formats.enums import DataSetType
 from wwt_data_formats.imageset import ImageSet
 from wwt_data_formats.place import Place
 
-from . import CxClient, ImageWwt, ImageStorage, _strip_nulls_in_place
+from . import CxClient
+from .data import (
+    ImageWwt,
+    ImageStorage,
+    SceneContent,
+    SceneImageLayer,
+    ScenePlace,
+    _strip_nulls_in_place,
+)
 
 __all__ = """
 AddImageRequest
 AddSceneRequest
 HandleClient
-SceneImageLayer
-SceneContent
-ScenePlace
 """.split()
 
 
 H2R = math.pi / 15
 D2R = math.pi / 180
-
-
-@dataclass_json
-@dataclass
-class ScenePlace:
-    ra_rad: float
-    dec_rad: float
-    zoom_deg: float
-    roll_rad: float
-
-
-@dataclass_json
-@dataclass
-class SceneImageLayer:
-    image_id: str
-    opacity: float
-
-
-@dataclass_json
-@dataclass
-class SceneContent:
-    image_layers: Optional[List[SceneImageLayer]]
 
 
 @dataclass_json
@@ -66,7 +49,6 @@ class AddImageRequest:
 @dataclass_json
 @dataclass
 class AddImageResponse:
-    error: bool
     id: str
     rel_url: str
 
@@ -83,7 +65,6 @@ class AddSceneRequest:
 @dataclass_json
 @dataclass
 class AddSceneResponse:
-    error: bool
     id: str
     rel_url: str
 
@@ -123,7 +104,9 @@ class HandleClient:
             http_method="POST",
             json=_strip_nulls_in_place(image.to_dict()),
         )
-        resp = AddImageResponse.schema().load(resp.json())
+        resp = resp.json()
+        resp.pop("error")
+        resp = AddImageResponse.schema().load(resp)
         return resp.id
 
     def add_image_from_set(self, imageset: ImageSet) -> str:
@@ -188,7 +171,9 @@ class HandleClient:
             http_method="POST",
             json=_strip_nulls_in_place(scene.to_dict()),
         )
-        resp = AddSceneResponse.schema().load(resp.json())
+        resp = resp.json()
+        resp.pop("error")
+        resp = AddSceneResponse.schema().load(resp)
         return resp.id
 
     def add_scene_from_place(self, place: Place) -> str:
